roaringbitmap migration to 64-bit
surprisingly easy. also, early stub of app class

+++ util/roaring/roaring_bitmap_test.go
@@ -8,7 +8,7 @@ func TestRoaringBitmap(t *testing.T) {
   r := NewRoaringBitset()
 
   for i := 0; i < 10_000; i++ {
-    error := r.InsertOne(int32(i))
+    error := r.InsertOne(uint64(i))
     if error != nil {
       t.Fatalf("Error inserting %d: [%v]", i, error)
     }
@@ -16,26 +16,26 @@ func TestRoaringBitmap(t *testing.T) {
 
 
   for i := 0; i < 10_000; i++ {
-    if !r.Has(int32(i)) {
+    if !r.Has(uint64(i)) {
       t.Fatalf("Expected %d to be in the bitmap", i)
     }
   }
 
   for i := 90_000; i < 100_000; i++ {
-    error := r.InsertOne(int32(i))
+    error := r.InsertOne(uint64(i))
     if error != nil {
       t.Fatalf("Error inserting %d: [%v]", i, error)
     }
   }
 
   for i := 90_000; i < 100_000; i++ {
-    if !r.Has(int32(i)) {
+    if !r.Has(uint64(i)) {
       t.Fatalf("Expected %d to be in the bitmap", i)
     }
   }
 
   for i := 10_000; i < 90_000; i++ {
-    if r.Has(int32(i)) {
+    if r.Has(uint64(i)) {
       t.Fatalf("Expected %d to not be in the bitmap", i)
     }
   }
@@ -45,6 +45,6 @@ func BenchmarkRoaringBitmapInsert(b *testing.B) {
   r := NewRoaringBitset()
 
   for i := 0; i < b.N; i++ {
-    r.InsertOne(int32(i))
+    r.InsertOne(uint64(i))
   }
 }
